organize as monorepo with lerna
