+++ package.json
@@ -1,7 +1,6 @@
 {
+  "private": true,
   "name": "unified-doc-dom",
-  "version": "0.1.7-beta.0",
-  "description": "DOM APIs for unified-doc",
   "license": "MIT",
   "homepage": "https://github.com/unified-doc/unified-doc-dom",
   "repository": "unified-doc/unified-doc-dom",
@@ -29,33 +28,33 @@
     "browser"
   ],
   "scripts": {
-    "bootstrap": "npm install",
-    "build": "microbundle",
-    "clean": "rm -rf dist",
-    "dev": "microbundle watch",
+    "bootstrap": "npm install; lerna bootstrap",
+    "build": "lerna exec --parallel microbundle",
+    "clean": "lerna exec --parallel 'rm -rf dist node_modules'",
+    "dev": "lerna exec --parallel 'microbundle watch'",
+    "link": "lerna link convert",
     "lint": "xo --fix; tsc",
-    "prepare": "npm run build",
-    "test": "echo 'probably too much effort to test.  PRs welcome :)'",
-    "test:run": "npm run test"
+    "test": "jest --watch packages",
+    "test:ci": "jest",
+    "prepare": "npm run clean; npm run build",
+    "publish": "lerna publish"
   },
-  "main": "dist/index.js",
-  "module": "dist/index.module.js",
-  "source": "index.js",
-  "typings": "types/index.d.ts",
   "dependencies": {
-    "rangy": "^1.3.0",
-    "resize-observer-polyfill": "^1.5.1"
+    "unified-doc-dom": "file:packages/unified-doc-dom"
   },
   "devDependencies": {
     "@babel/preset-env": "^7.10.3",
+    "@types/jest": "^26.0.9",
+    "babel-jest": "^26.2.2",
     "husky": "^4.2.5",
+    "jest": "^26.2.2",
     "microbundle": "^0.12.2",
     "typescript": "^3.9.5",
     "xo": "^0.32.0"
   },
   "husky": {
     "hooks": {
-      "pre-push": "npm run lint; npm run test:run; npm run clean; npm run build"
+      "pre-push": "npm prepare; npm run lint; npm run test:ci"
     }
   },
   "prettier": {
@@ -66,7 +65,8 @@
   },
   "xo": {
     "env": [
-      "browser"
+      "browser",
+      "jest"
     ],
     "prettier": true,
     "rules": {
