+++ packages/unified-doc-dom/lib/select-text.js
