+++ packages/unified-doc-dom/package.json
@@ -1,0 +1,43 @@
+{
+  "name": "unified-doc-dom",
+  "version": "1.0.0",
+  "description": "DOM APIs for unified-doc",
+  "license": "MIT",
+  "homepage": "https://github.com/unified-doc/unified-doc-dom",
+  "repository": "https://github.com/unified-doc/unified-doc-dom/tree/main/packages/unified-doc-dom",
+  "bugs": "https://github.com/unified-doc/unified-doc-dom/issues",
+  "author": "Chris Zhou <chrisrzhou@pm.me> (https://chrisrzhou.io)",
+  "keywords": [
+    "unified",
+    "doc",
+    "dom",
+    "document",
+    "file",
+    "content",
+    "unist",
+    "hast",
+    "parse",
+    "html",
+    "markdown",
+    "text",
+    "compile",
+    "export",
+    "annotate",
+    "highlight",
+    "nlp",
+    "search",
+    "browser"
+  ],
+  "main": "dist/index.js",
+  "module": "dist/index.module.js",
+  "source": "index.js",
+  "typings": "types/index.d.ts",
+  "files": [
+    "dist",
+    "types/index.d.ts"
+  ],
+  "dependencies": {
+    "rangy": "^1.3.0",
+    "resize-observer-polyfill": "^1.5.1"
+  }
+}
