+++ packages/unified-doc-dom/tests/highlight.test.js
@@ -1,0 +1,3 @@
+describe('highlight', () => {
+  it('should pass', () => {});
+});
