+++ packages/unified-doc-dom/tests/select-text.test.js
@@ -1,0 +1,3 @@
+describe('select-text', () => {
+  it('should pass', () => {});
+});
