+++ packages/unified-doc-dom/types/highlight.d.ts
@@ -9,23 +9,8 @@ export interface HighlightOptions {
   };
 }
 
-export interface SelectedText {
-  start: number;
-  end: number;
-  value: string;
-}
-
-export interface SelectTextOptions {
-  callback?: (selectedText: SelectedText) => void;
-}
-
 export function highlight(
   doc: HTMLElement,
   elementId: string,
   options?: HighlightOptions,
-): () => void;
-
-export function selectText(
-  doc: HTMLElement,
-  options?: SelectTextOptions,
 ): () => void;
