+++ packages/unified-doc-dom/types/index.d.ts
@@ -1,0 +1,2 @@
+export * from './highlight';
+export * from './select-text';
